Bunch of enhancements to monster pathfinding.
Monster won't try to walk out of bounds or over walls. Monster both checks this and the scent map does not mark walls.

Big levelup on the ScentMap debug overlay. Now changes colors based on how many turns ago the tile was marked and shades (dims) based on the distance from the player at the time.